IOS-108 added GalleryView, CarouselView
