+++ Datify-iOS-Core/Extensions/Array+extension.swift
@@ -9,4 +9,10 @@ import Foundation
 
 extension Array {
     var isNotEmpty: Bool { !self.isEmpty }
+
+    func chunked(into size: Int) -> [[Element]] {
+        return stride(from: 0, to: count, by: size).map {
+            Array(self[$0 ..< Swift.min($0 + size, count)])
+        }
+    }
 }
