+++ Datify-iOS-Core/Views/Main/Tabbar/Dating/DaringLikes/Carousel/SmallUserViewModel.swift
