+++ Datify-iOS-Core/Views/Main/Tabbar/Dating/DaringLikes/LikesView.swift
@@ -7,11 +7,11 @@
 
 import SwiftUI
 
-enum DisplayMode {
-    case gallery, carousel
-}
-
 struct LikesView: View {
+    enum DisplayMode {
+        case gallery, carousel
+    }
+
     @Environment (\.dismiss) private var dismiss
     @StateObject private var viewModel: LikesViewModel
     @State private var tag: LikeTage = .receivedLikes
@@ -19,7 +19,6 @@ struct LikesView: View {
     @State private var showFilters: Bool = false
     @State private var showInformationView: Bool = false
     @State private var blurRadius: CGFloat = 0
-    @State private var spacingGalleryCard: CGFloat = 6
 
     init(router: Router<AppRoute>) {
         _viewModel = StateObject(wrappedValue: LikesViewModel(router: router,
@@ -32,14 +31,26 @@ struct LikesView: View {
         NavigationStack {
             GeometryReader { geometry in
                 ZStack(alignment: .bottom) {
-                    VStack(alignment: .center) {
-                        categoriesFilter(size: geometry.size)
-                        switch displayMode {
-                        case .carousel: carouselUserLikes(size: geometry.size)
-                        case .gallery: gallerylUserLikes(size: geometry.size)
+                    VStack {
+                        if let currentUser = viewModel.currentUser {
+                            categoriesFilter(size: geometry.size)
+                            switch displayMode {
+                            case .carousel: CarouselView(size: geometry.size,
+                                                             currentUser: currentUser,
+                                                             likes: getLikesAndSelectedItem().likes,
+                                                             selectedItem: getLikesAndSelectedItem().selected,
+                                                             showInformationView: $showInformationView,
+                                                             blurRadius: $blurRadius)
+                            case .gallery: GallaryView(size: geometry.size,
+                                                           likes: chunkedArray(elements: 2),
+                                                           currentUser: currentUser,
+                                                           myLike: viewModel.myLikes)
+                            }
                         }
                     }
-                    .blur(radius: blurRadius)
+                    .sheetFilter(isPresented: $showFilters, blurRadius: $blurRadius, title: "Filter") {
+                        FilterView(sortOption: $viewModel.sortOption)
+                    }
                     InformationView(showView: $showInformationView,
                                     width: geometry.size.width,
                                     title: "Restore chat?",
@@ -58,9 +69,6 @@ struct LikesView: View {
                     blurRadius = newValue ? 10.0 : 0
                 }
             }
-            .sheetFilter(isPresented: $showFilters, blurRadius: $blurRadius) {
-                FilterView(sortOption: $viewModel.sortOption)
-            }
             .onAppear {
                 Task {
                     await viewModel.fecthData()
@@ -72,7 +80,6 @@ struct LikesView: View {
                 dtToolbarButton(placement: .topBarLeading, image: DtImage.backButton) {
                     dismiss()
                 }
-                // TODO: Replace with image from assets
                 dtToolbarButton(placement: .topBarTrailing, image: DtImage.likeFilter) {
                     showFilters.toggle()
                 }
@@ -112,87 +119,6 @@ struct LikesView: View {
     }
 
     @ViewBuilder
-    private func carouselUserLikes(size: CGSize) -> some View {
-        VStack {
-            Spacer()
-            BigUserCardView(selectedItem: getLikesAndSelectedItem().selected,
-                            size: size,
-                            showInformationView: $showInformationView,
-                            blurRadius: $blurRadius)
-            Spacer()
-            ScrollView(.horizontal, showsIndicators: false) {
-                HStack(spacing: 2) {
-                    HStack {Text("")}
-                        .frame(width: size.width / 2 - (size.width*0.07) / 2)
-                    ForEach(getLikesAndSelectedItem().likes) { like in
-                        if let currentUser = viewModel.currentUser {
-                            SmallUserCardView(like: like,
-                                              selectedItem: getLikesAndSelectedItem().selected,
-                                              currentUser: currentUser,
-                                              size: size)
-                        }
-                    }
-                    HStack {Text("")}
-                        .frame(width: size.width / 2 - (size.width*0.07) / 2)
-                }
-            }
-            .padding(.bottom)
-        }
-    }
-
-    @ViewBuilder
-    private func gallerylUserLikes(size: CGSize) -> some View {
-        VStack {
-            if getLikesAndSelectedItem().likes.isNotEmpty {
-                ScrollView(showsIndicators: false) {
-                    ForEach(chunkedArray(elements: 2).indices, id: \.self) { index in
-                        HStack(alignment: index == 0 ? .top  :  .bottom, spacing: spacingGalleryCard) {
-                            if let currentUser = viewModel.currentUser {
-                                MediumUserCardView(like: chunkedArray(elements: 2)[index][0],
-                                                   currentUser: currentUser,
-                                                   isCrop: isCrop(index: index,
-                                                                  indexNestedArray: 0,
-                                                                  count: chunkedArray(elements: 2).count),
-                                                   myLikes: viewModel.myLikes,
-                                                   size: size,
-                                                   spacing: spacingGalleryCard)
-                                .offset(y: isCrop(index: index,
-                                                  indexNestedArray: 0,
-                                                  count: chunkedArray(elements: 2).count) ? -size.height * 0.07 : 0)
-                                if chunkedArray(elements: 2)[index].count > 1 {
-                                    MediumUserCardView(like: chunkedArray(elements: 2)[index][1],
-                                                       currentUser: currentUser,
-                                                       isCrop: isCrop(index: index,
-                                                                      indexNestedArray: 1,
-                                                                      count: chunkedArray(elements: 2).count),
-                                                       myLikes: viewModel.myLikes,
-                                                       size: size,
-                                                       spacing: spacingGalleryCard)
-                                    .offset(y: isCrop(index: index,
-                                                      indexNestedArray: 1,
-                                                      count: chunkedArray(elements: 2).count) ? 0 : -size.height * 0.07)
-                                } else {
-                                    Rectangle()
-                                        .frame(height: size.height * 0.4)
-                                        .foregroundStyle(.clear)
-                                    Spacer()
-                                }
-                            }
-                        }
-                        .frame(width: size.width*0.92)
-                    }
-                }
-
-            } else {
-                NoLikesYetView(width: size.width * 0.92, height: size.height * 0.85)
-                DtButton(title: "Continue".localize(), style: .main) { }
-                .frame(width: size.width * 0.92)
-            }
-        }
-        .frame(width: size.width)
-    }
-
-    @ViewBuilder
     private func categoriesFilter(size: CGSize) -> some View {
         VStack {
             DtSegmentedPicker(selectedItem: $viewModel.categories, items: LikeTage.allCases) {
@@ -213,7 +139,7 @@ struct LikesView: View {
     }
 }
 
-extension LikesView {
+private extension LikesView {
     var filterSheetView: some View {
         NavigationStack {
             GeometryReader { geometry in
@@ -264,15 +190,6 @@ extension LikesView {
 
         let slope = (endValue - startValue) / (endHeight - startHeight)
         return startValue + slope * (height - startHeight)
-    }
-}
-
-// TODO: перенести в extention
-extension Array {
-    func chunked(into size: Int) -> [[Element]] {
-        return stride(from: 0, to: count, by: size).map {
-            Array(self[$0 ..< Swift.min($0 + size, count)])
-        }
     }
 }
